chore: data upload to algolia
